refactor: refactor ButtonBig and delete LoginPageBtn
