+++ rs-commerce/src/components/button-big/button-big.tsx
@@ -1,10 +1,16 @@
 import { ButtonBigProps } from 'types/types';
+import style from './style.module.css';
 
 function ButtonBig(props: ButtonBigProps): JSX.Element {
-  const { style, content, onClick } = props;
+  const { isActiveStyle, content, onClick } = props;
+
+  const btnStyle = (): string =>
+    isActiveStyle
+      ? `${style['login-page-btn']}`
+      : `${style['login-page-btn']} ${style['btn-not-active']}`;
 
   return (
-    <button type="button" className={style} onClick={onClick}>
+    <button type="button" className={btnStyle()} onClick={onClick}>
       {content}
     </button>
   );
